+++ internal/Panels/Viewport.js
@@ -1,5 +1,6 @@
 import classnames from 'classnames';
 import {forward, handle} from '@enact/core/handle';
+import {mapAndFilterChildren} from '@enact/core/util';
 import Spotlight from '@enact/spotlight';
 import Pause from '@enact/spotlight/Pause';
 import ViewManager, {shape} from '@enact/ui/ViewManager';
@@ -23,9 +24,9 @@ const PanelsStateContext = React.createContext({});
  * @private
  */
 const ViewportBase = class extends React.Component {
-	static displayName = 'Viewport'
+	static displayName = 'Viewport';
 
-	static contextType = SharedState
+	static contextType = SharedState;
 
 	static propTypes = /** @lends sandstone/Panels.Viewport.prototype */ {
 
@@ -44,6 +45,18 @@ const ViewportBase = class extends React.Component {
 		 * @type {Arranger}
 		 */
 		arranger: shape,
+
+		/**
+		 * Forwarded as `autoFocus` to each panel.
+		 *
+		 * If the `Panel` already has an `autoFocus` prop, it is maintained.
+		 *
+		 * Otherwise, if `autoFocus` is set, the value of this prop is added to the props. If it is
+		 * unset, 'default-element' is passed when navigating to a higher index.
+		 *
+		 * @type {String}
+		 */
+		autoFocus: PropTypes.string,
 
 		/**
 		 * Sets the hint string read when focusing the back button.
@@ -125,12 +138,12 @@ const ViewportBase = class extends React.Component {
 		onClose: PropTypes.func,
 
 		type: PropTypes.string
-	}
+	};
 
 	static defaultProps = {
 		index: 0,
 		noAnimation: false
-	}
+	};
 
 	constructor () {
 		super();
@@ -188,7 +201,7 @@ const ViewportBase = class extends React.Component {
 		}
 
 		return true;
-	}
+	};
 
 	removeTransitioningClass = () => {
 		if (this.node) {
@@ -196,45 +209,48 @@ const ViewportBase = class extends React.Component {
 		}
 
 		return true;
-	}
+	};
 
-	pause = () => this.paused.pause()
+	pause = () => this.paused.pause();
 
-	resume = () => this.paused.resume()
+	resume = () => this.paused.resume();
 
-	handle = handle.bind(this)
+	handle = handle.bind(this);
 
 	handleTransition = this.handle(
 		forward('onTransition'),
 		this.removeTransitioningClass,
 		this.resume
-	)
+	);
 
 	handleWillTransition = this.handle(
 		forward('onWillTransition'),
 		this.addTransitioningClass,
 		this.pause
-	)
+	);
 
-	mapChildren = (children, generateId) => React.Children.map(children, (child, index) => {
-		if (child) {
-			const {spotlightId = generateId(index, 'panel-container', Spotlight.remove)} = child.props;
-			const props = {
-				spotlightId,
-				'data-index': index
-			};
+	mapChildren = (children, generateId) => mapAndFilterChildren(children, (child, index) => {
+		const {spotlightId = generateId(index, 'panel-container', Spotlight.remove)} = child.props;
+		const props = {
+			spotlightId,
+			'data-index': index
+		};
 
-			if (child.props.autoFocus == null && this.state.direction === 'forward') {
+		// Respect Panel-configured autoFocus if it is set
+		if (child.props.autoFocus == null) {
+			if (this.props.autoFocus) {
+				// if not and Viewport-wide autoFocus is configured, use it
+				props.autoFocus = this.props.autoFocus;
+			} else if (this.state.direction === 'forward') {
+				// Otherwise, only set autofocus when moving forward
 				props.autoFocus = 'default-element';
 			}
-
-			return React.cloneElement(child, props);
-		} else {
-			return null;
 		}
-	})
 
-	getEnteringProp = (noAnimation) => noAnimation ? null : 'hideChildren'
+		return React.cloneElement(child, props);
+	});
+
+	getEnteringProp = (noAnimation) => noAnimation ? null : 'hideChildren';
 
 	render () {
 		const {