+++ samples/qa-a11y/src/views/ImageItem.js
@@ -2,6 +2,13 @@ import Heading from '@enact/sandstone/Heading';
 import ImageItem from '@enact/sandstone/ImageItem';
 import React from 'react';
 
+const SelectableImageItem = (props) => {
+	const [checked, setChecked] = React.useState(false);
+	const handleClick = () => setChecked(!checked);
+
+	return <ImageItem {...props} onClick={handleClick} selected={checked} />; // eslint-disable-line react/jsx-no-bind
+};
+
 const ImageItemView = () => (
 	<>
 		<ImageItem
@@ -11,7 +18,16 @@ const ImageItemView = () => (
 		>
 			ImageItem Caption
 		</ImageItem>
-		<Heading showLine>Aria-labled Items</Heading>
+		<Heading showLine>ImageItem with showSelection</Heading>
+		<SelectableImageItem
+			label="ImageItem label"
+			orientation="horizontal"
+			showSelection
+			src="http://via.placeholder.com/200x200/7ed31d/ffffff"
+		>
+			ImageItem Caption
+		</SelectableImageItem>
+		<Heading showLine>Aria-labelled ImageItem</Heading>
 		<ImageItem
 			aria-label="This is an image item"
 			label="ImageItem label"
@@ -20,6 +36,16 @@ const ImageItemView = () => (
 		>
 			ImageItem Caption
 		</ImageItem>
+		<Heading showLine>Aria-labelled ImageItem with showSelection</Heading>
+		<SelectableImageItem
+			aria-label="This is an image item"
+			label="ImageItem label"
+			orientation="horizontal"
+			showSelection
+			src="http://via.placeholder.com/200x200/7ed31d/ffffff"
+		>
+			ImageItem Caption
+		</SelectableImageItem>
 	</>
 );